+++ packages/cookie/CHANGELOG.md
@@ -6,6 +6,7 @@ This is the changelog for [`cookie`](https://github.com/remix-run/remix/tree/mai
 
 - BREAKING CHANGE: Rename `cookie.isSigned` to `cookie.signed`
 - Add `createCookie` function to create a new `Cookie` object
+- `CookieOptions` now extends `CookieProperties` so all cookie properties may be set in the `Cookie` constructor
 
 ## v0.2.0 (2025-11-04)